# docs/api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EduMark",
    "description": "\n基于大模型的中小学作业批改系统，提供以下核心能力：\n\n- **作业提交与批改**：上传作业图片，自动识别题目并进行 AI 批改\n- **知识库管理**：上传教学资料，构建知识库辅助批改\n- **用户管理**：学生/老师/管理员三种角色，权限分级\n- **模型管理**：配置视觉模型和语言模型\n- **配置管理**：个性化识别模式与增强选项\n",
    "contact": {
      "name": "EduMark Team",
      "email": "admin@edumark.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "认证管理"
        ],
        "summary": "用户注册",
        "description": "注册新用户账号，系统自动创建默认配置。\n\n- **account**：登录账号，需全局唯一\n- **password**：明文密码，服务端自动加密存储\n- **role**：角色，可选 `student`（学生）或 `teacher`（老师）",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              },
              "examples": {
                "注册学生用户": {
                  "summary": "创建一个学生账号",
                  "value": {
                    "account": "zhangsan",
                    "username": "张三",
                    "password": "123456",
                    "email": "zhangsan@example.com",
                    "role": "student"
                  }
                },
                "注册老师用户": {
                  "summary": "创建一个老师账号",
                  "value": {
                    "account": "teacher_wang",
                    "username": "王老师",
                    "password": "123456",
                    "email": "wang@example.com",
                    "role": "teacher"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "注册成功，返回用户信息",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_UserOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "认证管理"
        ],
        "summary": "用户登录",
        "description": "使用账号密码登录，验证成功后返回 JWT 令牌。\n\n- **access_token**：访问令牌，用于后续接口鉴权，有效期较短\n- **refresh_token**：刷新令牌，用于获取新的 access_token，有效期较长",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginSchema"
              },
              "examples": {
                "学生登录": {
                  "summary": "学生账号登录",
                  "value": {
                    "account": "zhangsan",
                    "password": "123456"
                  }
                },
                "老师登录": {
                  "summary": "老师账号登录",
                  "value": {
                    "account": "teacher_wang",
                    "password": "123456"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "登录成功，返回 access_token 和 refresh_token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_TokenResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "认证管理"
        ],
        "summary": "刷新令牌",
        "description": "使用 refresh_token 换取新的 access_token。\n\n- 每次刷新后，旧的 refresh_token 依然有效\n- 返回新的 access_token 和 refresh_token 对",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshSchema"
              },
              "examples": {
                "刷新令牌": {
                  "summary": "使用 refresh_token 换取新令牌",
                  "value": {
                    "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "刷新成功，返回新的令牌对",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_TokenResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "认证管理"
        ],
        "summary": "当前用户信息",
        "description": "获取当前登录用户的详细信息，需要在请求头中携带有效的 access_token。",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "返回当前用户信息",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_UserOut_"
                }
              }
            }
          },
          "401": {
            "description": "未认证或 token 无效"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/v1/config/me": {
      "get": {
        "tags": [
          "配置管理"
        ],
        "summary": "获取配置",
        "description": "获取当前登录用户的个性化配置。",
        "operationId": "get_my_config_api_v1_config_me_get",
        "responses": {
          "200": {
            "description": "查询成功，返回用户配置",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_ConfigOut_"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "配置管理"
        ],
        "summary": "修改配置",
        "description": "修改当前用户的个性化配置，包含模型校验。\n\n- **rec_mode**：`aliyun`（阿里云）或 `bailian`（百炼）\n- **vl_model**：视觉模型名称，当 rec_mode 为 bailian 时必填\n- **gl_model**：语言模型名称\n- **enable_enhance**：是否启用增强功能\n- **enable_knowledge**：是否启用知识库检索",
        "operationId": "update_my_config_api_v1_config_me_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigUpdate"
              },
              "examples": {
                "阿里云模式": {
                  "summary": "使用阿里云识别模式，开启增强和知识库",
                  "value": {
                    "rec_mode": "aliyun",
                    "enable_enhance": true,
                    "enable_knowledge": true,
                    "gl_model": "qwen-max"
                  }
                },
                "百炼模式": {
                  "summary": "使用百炼识别模式，需指定视觉模型",
                  "value": {
                    "rec_mode": "bailian",
                    "enable_enhance": true,
                    "enable_knowledge": true,
                    "vl_model": "qwen-vl-max",
                    "gl_model": "qwen-max"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "更新成功，返回配置信息",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_ConfigOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfigOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "配置记录 ID"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "所属用户 ID"
          },
          "rec_mode": {
            "type": "string",
            "title": "Rec Mode",
            "description": "识别模式"
          },
          "enable_enhance": {
            "type": "boolean",
            "title": "Enable Enhance",
            "description": "是否启用增强功能"
          },
          "enable_knowledge": {
            "type": "boolean",
            "title": "Enable Knowledge",
            "description": "是否启用知识库"
          },
          "vl_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vl Model",
            "description": "视觉模型名称"
          },
          "gl_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gl Model",
            "description": "语言模型名称"
          },
          "create_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Create Time",
            "description": "创建时间"
          },
          "update_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Update Time",
            "description": "最后更新时间"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "rec_mode",
          "enable_enhance",
          "enable_knowledge"
        ],
        "title": "ConfigOut",
        "example": {
          "create_time": "2026-06-01T08:00:00",
          "enable_enhance": true,
          "enable_knowledge": true,
          "gl_model": "qwen-max",
          "id": 1,
          "rec_mode": "aliyun",
          "update_time": "2026-06-04T12:00:00",
          "user_id": 1,
          "vl_model": "qwen-vl-max"
        }
      },
      "ConfigUpdate": {
        "properties": {
          "rec_mode": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(aliyun|bailian)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rec Mode",
            "description": "识别模式：aliyun=阿里云，bailian=百炼"
          },
          "enable_enhance": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enable Enhance",
            "description": "是否启用增强功能"
          },
          "enable_knowledge": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enable Knowledge",
            "description": "是否启用知识库"
          },
          "vl_model": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Vl Model",
            "description": "视觉模型名称"
          },
          "gl_model": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Gl Model",
            "description": "语言模型名称"
          }
        },
        "type": "object",
        "title": "ConfigUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginSchema": {
        "properties": {
          "account": {
            "type": "string",
            "maxLength": 64,
            "minLength": 3,
            "title": "Account",
            "description": "用户账号，3~64 个字符"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 6,
            "title": "Password",
            "description": "用户密码，6~128 个字符"
          }
        },
        "type": "object",
        "required": [
          "account",
          "password"
        ],
        "title": "LoginSchema"
      },
...
      "RefreshSchema": {
        "properties": {
          "token": {
            "type": "string",
            "minLength": 1,
            "title": "Token",
            "description": "刷新令牌（refresh_token）"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "RefreshSchema"
      },
...
      "Result_ConfigOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码，1=成功，-1=失败"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "提示信息，成功时为 'success'，失败时为错误描述"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfigOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "响应数据，成功时返回业务数据，失败时为 null"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Result[ConfigOut]",
        "example": {
          "code": 1,
          "message": "success"
        }
      },
...
      "Result_TokenResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码，1=成功，-1=失败"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "提示信息，成功时为 'success'，失败时为错误描述"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TokenResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "响应数据，成功时返回业务数据，失败时为 null"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Result[TokenResponse]",
        "example": {
          "code": 1,
          "message": "success"
        }
      },
      "Result_UserOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码，1=成功，-1=失败"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "提示信息，成功时为 'success'，失败时为错误描述"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "响应数据，成功时返回业务数据，失败时为 null"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Result[UserOut]",
        "example": {
          "code": 1,
          "message": "success"
        }
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token",
            "description": "访问令牌，有效期 120 分钟"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token",
            "description": "刷新令牌，有效期 7 天"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "description": "令牌类型，固定为 bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "token_type"
        ],
        "title": "TokenResponse",
        "example": {
          "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
          "refresh_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
          "token_type": "bearer"
        }
      },
      "UserCreate": {
        "properties": {
          "account": {
            "type": "string",
            "maxLength": 64,
            "minLength": 3,
            "title": "Account",
            "description": "用户账号，3~64 个字符"
          },
          "username": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Username",
            "description": "用户昵称，1~64 个字符"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 6,
            "title": "Password",
            "description": "用户密码，6~128 个字符"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "description": "电子邮箱，可选"
          },
          "role": {
            "type": "string",
            "pattern": "^(student|teacher)$",
            "title": "Role",
            "description": "用户角色：student=学生，teacher=老师",
            "default": "student"
          }
        },
        "type": "object",
        "required": [
          "account",
          "username",
          "password"
        ],
        "title": "UserCreate"
      },
      "UserOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "用户 ID"
          },
          "account": {
            "type": "string",
            "title": "Account",
            "description": "用户账号"
          },
          "username": {
            "type": "string",
            "title": "Username",
            "description": "用户昵称"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "description": "电子邮箱"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "description": "用户角色：student 或 teacher"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "是否启用"
          },
          "create_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Create Time",
            "description": "创建时间"
          },
          "update_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Update Time",
            "description": "最后更新时间"
          }
        },
        "type": "object",
        "required": [
          "id",
          "account",
          "username",
          "role",
          "is_active"
        ],
        "title": "UserOut",
        "example": {
          "account": "zhangsan",
          "create_time": "2026-06-01T08:00:00",
          "email": "zhangsan@example.com",
          "id": 1,
          "is_active": true,
          "role": "student",
          "update_time": "2026-06-04T12:00:00",
          "username": "张三"
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  },
...
}